Got all up and running
The Bluetooth data transferred from the ESP is now stored in the dictionary when you run the update method the code isn't hard yet it is hardcoded as that is the simpelest method. this code can be modified but it will take a while since i had to manually assign each value to a variable. variables are stored in the data[] dictionary in the python101.py script.