All issues resolved
fixed threads locking, main thread now starts in python101.py and is transferred to the Macro's file in the first few lines.

data now comes through as all 0 without a device instead of crashing the script.

sockets have been moved to the bottom of the file to make sure the background threads are started properly

any issues be sure to ask me.